+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+next
+====
+
+* Added EncryptionMasterKeyBase64 parameter
+* Deprecated EncryptionMasterKey parameter
+
 4.0.0 / 2019-05-31
 ==================
 * This release modifies the entire repo to respect the go linter. This is a significant API breaking change and will likely require you to correct references to the names that were changed in your code. All future releases will respect the linter. A summary of the changes: